requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow
keras
# opencv-python
h5py
imgaug
# IPython[all]
tensorflow-gpu
pip
setuptools
# imantics
psutil
# python-socketio[client]